refactor: playlistItem
