+++ src/components/playlists/View.tsx
@@ -1,4 +1,4 @@
-import React, { ReactNode, useRef, useState, useEffect } from 'react';
+import React, { useRef, useState, useEffect } from 'react';
 import { IconButton, Divider, Text, TouchableRipple } from 'react-native-paper';
 import { DrawerActions, useNavigation } from '@react-navigation/native';
 import {
@@ -30,6 +30,7 @@ import TimerButton from './TimerButton';
 import appStore from '@stores/appStore';
 import logger from '@utils/Logger';
 import useTheme from '@hooks/useTheme';
+import PlaylistItem from './PlaylistItem';
 
 const useRenderDrawerItem = () => {
   const navigation = useNavigation();
@@ -52,61 +53,6 @@ const useRenderDrawerItem = () => {
         </View>
       </View>
     </TouchableRipple>
-  );
-};
-
-const DefaultIcon = (
-  item: NoxMedia.Playlist,
-  deleteCallback: (id: string) => void
-) => {
-  const playerStyle = useNoxSetting(state => state.playerStyle);
-
-  return (
-    <IconButton
-      icon="close"
-      onPress={() => deleteCallback(item.id)}
-      size={25}
-      iconColor={playerStyle.colors.primary}
-    />
-  );
-};
-
-interface PlaylistItemProps {
-  item: NoxMedia.Playlist;
-  icon?: ReactNode;
-  confirmOnDelete?: (id: string) => void;
-  leadColor?: string;
-}
-const PlaylistItem = ({
-  item,
-  icon,
-  confirmOnDelete = () => undefined,
-  leadColor,
-}: PlaylistItemProps) => {
-  const currentPlayingList = useNoxSetting(state => state.currentPlayingList);
-
-  if (!item) return <></>;
-  return (
-    <View style={styles.playlistItemContainer}>
-      <View style={{ backgroundColor: leadColor, width: 15 }}></View>
-      <View style={styles.playlistItemTextContainer}>
-        <Text
-          variant="bodyLarge"
-          style={[
-            {
-              fontWeight:
-                currentPlayingList.id === item?.id ? 'bold' : undefined,
-              paddingHorizontal: 10,
-            },
-          ]}
-        >
-          {item.title}
-        </Text>
-      </View>
-      <View style={styles.playlistItemIconContainer}>
-        {icon ? icon : DefaultIcon(item, () => confirmOnDelete(item.id))}
-      </View>
-    </View>
   );
 };
 
@@ -335,13 +281,6 @@ const styles = StyleSheet.create({
   topPadding: {
     height: 10,
   },
-  playlistItemContainer: {
-    flexDirection: 'row',
-  },
-  playlistItemTextContainer: {
-    flex: 4,
-    justifyContent: 'center',
-  },
   addPlaylistButtonContainer: {
     height: 50,
     alignContent: 'center',
@@ -367,5 +306,4 @@ const styles = StyleSheet.create({
   },
   drawerItemContainer: { flexDirection: 'row' },
   drawerItemTextContainer: { justifyContent: 'center' },
-  playlistItemIconContainer: { alignItems: 'flex-end' },
 });
